update trial banner with intro offer + my-plan
