+++ packages/data-stores/src/plans/index.ts
@@ -20,7 +20,7 @@ export type {
 } from './types';
 
 export { default as usePricedAPISitePlans } from './queries/use-priced-api-site-plans';
-export { default as useIntroOffers } from './queries/use-intro-offers';
+export { default as useIntroOffers } from './hooks/use-intro-offers';
 
 // plansSlugs is a list with the identifiers for each plan and they are agnostic of billing period; eg: 'free', 'personal'
 // plansSlugs is also used to construct the route that accepts plan slugs like '/free', '/personal', '/business'
