+++ packages/data-stores/src/plans/queries/use-priced-api-site-plans.ts
@@ -16,6 +16,7 @@ interface Props {
 }
 
 /**
+ * Get all plans for a site, indexed by product_slug
  * - Plans from `/sites/[siteId]/plans`, unlike `/plans`, are returned indexed by product_id, and do not inlcude that in the plan's payload.
  * - UI works with product/plan slugs everywhere, so returned index is transformed to be keyed by product_slug
  */
